requirements.txt
google-generativeai==0.5.2
streamlit==1.33.0
pandas==2.2.2
pysqlite3-binary
chromadb==0.5.0
geopy
timezonefinder
emoji
Faker